add file, activity and library lists, with versions
Generates sorted lists of activity and library components by
descending the filesystem trees and reading the info files.

Generates a file list with the build number removed.

These are for build to build comparison and to be able to rapidly answer
support questions like "which activity version is in that build?".