sd_card_image: reorganise output
The rest of olpc-os-builder works under the assumption that everything
before the first "." in the output image name is user-controlled.
sd_card_image was an exception, but that is no longer the case.

The user now controls the output file extension (e.g. .zd, .zd4, .zd8) via
the config, rather than the old naming scheme of (e.g.) os4-4g.zd.

Also, the raw disk images are no longer useful on a day-to-day basis, so they
are now deleted by default.