11.2.0 config: switch to jffs2 on XO-1
We have 2 key problems with ubifs, which have only really come to light
in the last few days:

1. Space usage. We already knew that we grew 10-15% over jffs2, plus
the partitions overhead, but I now realise that the boot partition size
calculated earlier is far too small, as we ship both signed and unsigned
kernel/actrd/runrd. Once we fix the partition size, we really run low
on space.

2. Boot problems
We have had a couple of reports of unbootable XO-1s which need to be
looked into.

Disable ubifs for this release.