Forcefully remove mesa-dri-drivers (#11036)
In F14 mesa-dri-drivers was required as a dependency of mesa-libGL, but
since the version 7.10-0.10 (F15) it doesn't require it anymore and lets
comps do it.

Mesa-dri-drivers is useless for XO hardware, and it takes 41.83mb in the
F14 image. Remove this package is safe, the next versions like F15+ won't
be included automatically. We'll be able to remove this hack as soon
as we move to F15+.