repos: consider arch when building exclude list
This avoids the situation where a i686 xulrunner RPM causes xulrunner
to be added to the exclude list on ARM. As though there is no
xulrunner ARM RPM present in the OLPC repositories, this resulted in
xulrunner not being able to be found in any repository.

Fixed by only adding arch-relevant packages to the exclude list.