Store NetworkManager configs in /etc/NetworkManager (#9789)
Now that NM stores network configuration info at the system level,
we must take measures so that such configuration is retained over
upgrades.

Handling this in /etc/sysconfig/network-scripts would be a little
tricky/messy as that directory mixes code with data.

Disable the ifcfg-rh plugin so that configuration gets stored
in the pure-configuration directory at /etc/NetworkManager/system-settings
and put this path in statetab.