zhashfs: Use FIEMAP to determine used areas of the file (#10850)
Under F17, something (e2fsprogs?) is punching holes in the output disk
image, removing the 'fill' marker that had been put in place.

Switch to using FIEMAP to determine which blocks of the sparse disk image
are used. This also results in slightly smaller images when compared to
the previous approach.