sd_card_image: Calculate disk image size automatically
By default, calculate the output size of the resultant disk image
based on the amount of data needed to be stored, plus some free space.

The partition and filesystem will now be grown on first boot to fill the
media by other parts of the system.

The manual sizes can still be specified in the configuration if needed,
but the idea here is that we no longer need to produce disk images of
multiple sizes.