base: random-seed handling fixes
We put random-seed in persistent statetab, and I think that
makes more sense than rwtab. (Before this commit it was included in both.)

https://bugzilla.redhat.com/show_bug.cgi?id=808907 describes how
systemd's random seed handling and readonly-root aren't working well
together yet. Apply a solution to the bootup issue, and disable the
random-seed shutdown script which is not needed.