base: don't strip kernel modules
It looks like the strip invocation used here strips all symbols, even
when you ask for it to strip a specific section.

Exclude kernel modules from this so that they can still be loaded.

Hopefully this will be the last oddity caused by minidebuginfo-stripping.
If it causes much more trouble we should consider not doing this strip
at all.