Always ship kernel/initramfs in zip files
We now ship the kernel/initramfs in zip files only.
This reduces duplication for when images get signed, and it allows
us to ship a simplified olpc.fth.

This implements the plan here:
http://lists.laptop.org/pipermail/devel/2012-March/034439.html

This didn't work on first attempt because we hit a firmware bug
on XO-1.75 (#12107). Now that we have a fix, and a workaround for
old firmware versions, give this change another spin.