zhashfs: correct output when last block is not occupied (#12402)
zhashfs was writing the eblocks: count line based on the file size of
the input. However, this assumes that the last block of the input file
had data in it, which is not necessarily true. Resultant .zd images
flashed OK but caused OFW to produce a warning.

  WARNING:  The file said highest block 15911 but wrote only as high as block 15903

Adjust the calculation to look at the last occupied data block to fix
this.

Test case, last block used is in the middle of the file:
  rm -f testfile
  dd if=/dev/zero of=testfile count=0 seek=128M bs=1
  dd if=/dev/zero of=testfile count=1 seek=64M bs=1 conv=notrunc
  ./bin/zhashfs 0x20000 sha256 testfile testfile.zsp testfile.zd

The output used to trigger the OFW warning, now it does not.

Test case, last block used is at the end of the file:
  rm -f testfile
  dd if=/dev/zero of=testfile count=0 seek=128M bs=1
  dd if=/dev/zero of=testfile count=1 seek=134217727 bs=1
  ./bin/zhashfs 0x20000 sha256 testfile testfile.zsp testfile.zd

The output here is unchanged by this patch - good.