base: create compat symlink for bootfw too
Spotted by German in Nicaragua, an XO-1 running Q2E45 doesn't
auto-upgrade the firmware when 13.2.0 is installed. This is because
it looks for bootfw.zip whereas the installed file is named bootfw0.zip.

This breaks the rule that post-release commits should not change
the default build output. This one does, because it adds a symlink.
Given the current situation and the simplicity of this fix,
we'll break the rules.