pass --cache-only to scripts
The --cache-only flag was not being passed to scripts through the
environment variables.

This was most evident in
modules/sugar_activity_group/kspost.60.nochroot.activities.py which
re-read the activity group on every build even if --cache-only was set.

The cause was make_environment called by __init__ in OsBuilder, at a
time when self.cacheonly had not yet inherited the command line options.

The changes are to:

- add cacheonly on __init__ method arguments,

- store cacheonly in self.cacheonly before calling make_environment,

- remove cacheonly from OsBuilder.build method arguments.