Revert "base: configure avahi to avoid SO_REUSEPORT"
This reverts commit b9746dcbc13d8d6c7e2227edf0abc75183476e66, restoring
avahi-daemon to the default configuration which requires SO_REUSEPORT.

Our kernel now has a backport of support for SO_REUSEPORT, as of git hash
5196e01 on 2014-09-16.  This kernel is available in the OLPC dropbox:

http://dev.laptop.org/~kernels/public_rpms/f20-xo4/

SO_REUSEPORT is also required by telepathy-salut for preparing the
multicast transport when an activity is shared.  The symptom is missing
shared activities in network neighbourhood.

See gibber-multicast-transport.c which calls setsockopt with
SO_REUSEPORT if it is defined.