base: hack default route when using ad-hoc
On Fedora 20, when an ad-hoc network is connected, Sugar buddy icons and shared
activities are not shown in neighbourhood.  Activity logs contained "No
active connection available."  Salut logs contained "no connections, and
timed out."

It was discovered that configuring a default route avoided the problem,
though the mechanism by which this works is not known.

Add a Network Manager dispatcher script which creates a default route to
169.254.1.1 on eth0 up with matching connection ID and IP address.